ソースを整理しつつWindows上で動作

+++ src/chisa/donut/object/Object.cpp
@@ -26,7 +26,7 @@ namespace donut {
 static const std::string TAG("Object");
 static const std::string INDEX_STR[]={"0","1","2","3","4","5","6","7","8","9","10","11","12","13","14","15","16","17","18","19","20","21","22","23","24","25","26","27","28","29","30","31","32","33","34","35","36","37","38","39","40","41","42","43","44","45","46","47","48","49","50"};
 static std::string toName( const int& id){
-	return id >= 0 && id < (sizeof(INDEX_STR)/sizeof(std::string)) ? INDEX_STR[id] : util::format("%d", id);
+	return id >= 0 && static_cast<unsigned int>(id) < (sizeof(INDEX_STR)/sizeof(std::string)) ? INDEX_STR[id] : util::format("%d", id);
 }
